backend/api/requirements.txt
djangorestframework==3.13.1
bs4==0.0.1
requests==2.27.1
django-cors-headers==3.13.0